# swagger/swagger.yaml
basePath: /api/
consumes:
- application/json
definitions:
  AodAPIModel:
    properties:
      date:
        description: The date when this author was the author of the day
        example: "2021-06-12T00:00:00Z"
        type: string
        x-go-name: Date
      id:
        description: The author's id
        example: 24952
        format: int64
        type: integer
        x-go-name: Id
      name:
        description: The name of the author
        example: Muhammad Ali
        type: string
        x-go-name: Name
    type: object
    x-go-package: github.com/Skjaldbaka17/quotes-api/structs
  AuthorAPIModel:
    properties:
      count:
        description: The popularity index of the author
        example: 1111
        format: int64
        type: integer
        x-go-name: Count
      hasIcelandicQuotes:
        description: Whether or not this author has some icelandic quotes
        example: true
        type: boolean
        x-go-name: HasIcelandicQuotes
      id:
        description: The author's id
        example: 24952
        format: int64
        type: integer
        uniqueItems: true
        x-go-name: Id
      name:
        description: Name of the author
        example: Muhammad Ali
        type: string
        x-go-name: Name
      nrOfEnglishQuotes:
        description: How many quotes in English this author has
        example: 78
        format: int64
        type: integer
        x-go-name: NrOfEnglishQuotes
      nrOfIcelandicQuotes:
        description: How many quotes in Icelandic this author has
        example: 6
        format: int64
        type: integer
        x-go-name: NrOfIcelandicQuotes
    type: object
    x-go-package: github.com/Skjaldbaka17/quotes-api/structs
  OfTheDayModel:
    properties:
      date:
        description: The date when this author / quote should be 'of the day'
        example: "2020-06-12"
        type: string
        x-go-name: Date
      id:
        description: The id of the author / quote
        example: 1
        format: int64
        type: integer
        x-go-name: Id
      language:
        default: English
        description: The language of this author / quote
        example: icelandic
        type: string
        x-go-name: Language
    type: object
    x-go-name: ofTheDayModel
    x-go-package: github.com/Skjaldbaka17/quotes-api/swagger
  OrderConfiguration:
    properties:
      maximum:
        description: |-
          Where to end the ordering (if empty it ends at the logical end, for example end at the highest quoteid for quoteid ascending order).
          Note this key is always a string.
        example: "11"
        type: string
        x-go-name: Maximum
      minimum:
        description: |-
          Where to start the ordering (if empty it starts from beginning, for example start at 1 for quoteid ascending order).
          Note this key is always a string.
        example: "10"
        type: string
        x-go-name: Minimum
      orderBy:
        description: What to order by, 'quoteId', 'popularity' or 'length'
        example: popularity
        type: string
        x-go-name: OrderBy
      reverse:
        description: Whether to order the list in reverse or not (true is Descending
          and false is Ascending, false is default)
        example: true
        type: boolean
        x-go-name: Reverse
    type: object
    x-go-name: orderConfigListQuotesModel
    x-go-package: github.com/Skjaldbaka17/quotes-api/swagger
...
  SearchViewAPIModel:
    properties:
      authorId:
        description: The author's id
        example: 24952
        format: int64
        type: integer
        uniqueItems: true
        x-go-name: AuthorId
      isIcelandic:
        description: Whether or not this quote is in Icelandic or not
        example: false
        type: boolean
        x-go-name: IsIcelandic
      name:
        description: Name of author
        example: Muhammad Ali
        type: string
        x-go-name: Name
      quote:
        description: The quote
        example: Float like a butterfly, sting like a bee.
        type: string
        x-go-name: Quote
      quoteId:
        description: The quote's id
        example: 582676
        format: int64
        type: integer
        uniqueItems: true
        x-go-name: QuoteId
    type: object
    x-go-package: github.com/Skjaldbaka17/quotes-api/structs
...
host: quotel-api.com
info:
  contact:
    email: skjaldbaka17@gmail.com
    name: Þórður Ágústsson
  description: |-
    Quotel API offers a complete feature rich REST API access to all your quotes needs. This is the documentation of our quotes API.
    To use it you first need to sign up here https://somethingsomethingdarkside.com to get an authentication api key.
  title: quotes-api.
  version: 1.0.0
paths:
  /authors:
    post:
      description: Get the authors by their ids
      operationId: GetAuthors
      parameters:
      - description: The structure of the request for getting authors by their ids
        in: body
        name: Body
        required: true
        schema:
          properties:
            apiKey:
              description: The api-key you use to access the api
              example: 91fd6d19-2c32-4081-8729-4d9786d43b95
              type: string
              x-go-name: ApiKey
            ids:
              description: A list of the authors's ids that you want
              example:
              - 24952
              - 19161
              items:
                format: int64
                type: integer
              type: array
              x-go-name: Ids
          required:
          - ids
          - apiKey
          type: object
      responses:
        "200":
          $ref: '#/responses/authorsResponse'
        "400":
          $ref: '#/responses/incorrectBodyStructureResponse'
        "500":
          $ref: '#/responses/internalServerErrorResponse'
      tags:
      - AUTHORS
  /authors/aod:
    post:
      description: Gets the author of the day
      operationId: GetAuthorOfTheDay
      parameters:
      - description: The structure of the request for getting the author / quote of
          the day
        in: body
        name: Body
        schema:
          properties:
            apiKey:
              description: The api-key you use to access the api
              example: 91fd6d19-2c32-4081-8729-4d9786d43b95
              type: string
              x-go-name: ApiKey
            language:
              default: English
              description: Get the author / quote of the day for the given language
                ("icelandic" or "english")
              example: English
              type: string
              x-go-name: Language
          required:
          - apiKey
          type: object
      responses:
        "200":
          $ref: '#/responses/aodResponse'
        "400":
          $ref: '#/responses/incorrectBodyStructureResponse'
        "500":
          $ref: '#/responses/internalServerErrorResponse'
      tags:
      - AUTHORS
  /authors/aod/history:
    post:
      description: Gets the history for the authors of the day
      operationId: GetAODHistory
      parameters:
      - description: The structure of the request for getting the history of AODs
          / QODs
        in: body
        name: Body
        schema:
          items:
            properties:
              apiKey:
                description: The api-key you use to access the api
                example: 91fd6d19-2c32-4081-8729-4d9786d43b95
                type: string
                x-go-name: ApiKey
              language:
                default: English
                description: Get the history of the AODS / QODs for the given language
                  ("icelandic" or "english")
                example: icelandic
                type: string
                x-go-name: Language
              minimum:
                description: The earliest date to return. All authors / quotes between
                  minimum and today will be returned.
                example: "2020-12-21"
                type: string
                x-go-name: Minimum
            required:
            - apiKey
            type: object
          type: array
      responses:
        "200":
          $ref: '#/responses/aodHistoryResponse'
        "400":
          $ref: '#/responses/incorrectBodyStructureResponse'
        "500":
          $ref: '#/responses/internalServerErrorResponse'
      tags:
      - AUTHORS
  /authors/aod/new:
    post:
      description: sets the author of the day for the given dates
      operationId: SetAuthorOfTheDay
      parameters:
      - description: The structure of the request for setting AODs
        in: body
        name: Body
        schema:
          items:
            properties:
              aods:
                items:
                  $ref: '#/definitions/OfTheDayModel'
                type: array
                x-go-name: Aods
              apiKey:
                description: The api-key you use to access the api
                example: 91fd6d19-2c32-4081-8729-4d9786d43b95
                type: string
                x-go-name: ApiKey
            required:
            - apiKey
            type: object
          type: array
      responses:
        "200":
          $ref: '#/responses/successResponse'
        "400":
          $ref: '#/responses/incorrectBodyStructureResponse'
        "500":
          $ref: '#/responses/internalServerErrorResponse'
      tags:
      - AUTHORS
  /authors/list:
    post:
      description: Get a list of authors according to some ordering / parameters
      operationId: ListAuthors
      parameters:
      - description: The structure of the request for getting a list of authors
        in: body
        name: Body
        schema:
          properties:
            apiKey:
              description: The api-key you use to access the api
              example: 91fd6d19-2c32-4081-8729-4d9786d43b95
              type: string
              x-go-name: ApiKey
            language:
              description: |-
                Only return authors that have quotes in the given language ("english" or "icelandic") if left empty then no constraint
                is set on the quotes' language. Note if ordering by nrOfQuotes if this parameter is set then only the amount of
                quotes the author has in the given language counts towards the final ordering.
              example: English
              type: string
              x-go-name: Language
            orderConfig:
              $ref: '#/definitions/OrderConfiguration'
            page:
              description: Response is paged. This parameter controls the page you
                are asking for, starts with 0.
              example: 0
              format: int64
              minimum: 0
              type: integer
              x-go-name: Page
            pageSize:
              default: 25
              description: Response is paged. This parameter controls the number of
                Authors to be returned on each "page"
              example: 30
              format: int64
              maximum: 200
              minimum: 1
              type: integer
              x-go-name: PageSize
          required:
          - apiKey
          type: object
      responses:
        "200":
          $ref: '#/responses/authorsResponse'
        "400":
          $ref: '#/responses/incorrectBodyStructureResponse'
        "500":
          $ref: '#/responses/internalServerErrorResponse'
      tags:
      - AUTHORS
  /authors/random:
    post:
      description: Get a random Author, and some of his quotes, according to the given
        parameters
      operationId: GetRandomAuthor
      parameters:
      - description: The structure of the request for getting a random author
        in: body
        name: Body
        schema:
          properties:
            apiKey:
              description: The api-key you use to access the api
              example: 91fd6d19-2c32-4081-8729-4d9786d43b95
              type: string
              x-go-name: ApiKey
            language:
              description: |-
                The random author must have quotes in the given language ("english" or "icelandic") if left empty then no
                constraint on language is set
              example: English
              type: string
              x-go-name: Language
            maxQuotes:
              default: 1
              description: How many quotes, maximum, to be returned from this author
              example: 10
              format: int64
              maximum: 50
              type: integer
              x-go-name: MaxQuotes
          required:
          - apiKey
          type: object
      responses:
        "200":
          $ref: '#/responses/searchViewsResponse'
        "400":
          $ref: '#/responses/incorrectBodyStructureResponse'
        "500":
          $ref: '#/responses/internalServerErrorResponse'
      tags:
      - AUTHORS
...
produces:
- application/json
responses:
  aodHistoryResponse:
    description: Data structure representing the response for the history of AODs
    schema:
      items:
        $ref: '#/definitions/AodAPIModel'
      type: array
  aodResponse:
    description: Data structure representing the response for the author of the day
    schema:
      $ref: '#/definitions/AodAPIModel'
  authorsResponse:
    description: Data structure representing the response for authors
    schema:
      items:
        $ref: '#/definitions/AuthorAPIModel'
      type: array
  incorrectBodyStructureResponse:
    description: Data structure representing the error response to a wrongly structured
      request body
    schema:
      properties:
        message:
          description: The error message
          example: request body is not structured correctly.
          type: string
          x-go-name: Message
      type: object
...
  internalServerErrorResponse:
    description: Data structure representing the error response to an internal server
      error
    schema:
      properties:
        message:
          description: The error message
          example: Please try again later.
          type: string
          x-go-name: Message
      type: object
...
  searchViewsResponse:
    description: Data structure representing the response for quotes
    schema:
      items:
        $ref: '#/definitions/SearchViewAPIModel'
      type: array
  successResponse:
    description: ""
    schema:
      properties:
        message:
          example: This request was a success
          type: string
          x-go-name: Message
        statusCode:
          description: HTTP status code
          example: 200
          format: int64
          type: integer
          x-go-name: StatusCode
      type: object
...
schemes:
- http
securityDefinitions:
  apikey:
    description: To use the api you need an API key. The basic API key lets you make
      up to 100 requests an hour.
    in: body
    name: apiKey
    type: apiKey
swagger: "2.0"
...